requirements.txt
alembic==1.8.1
anyio==3.6.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.0
greenlet==1.1.3
h11==0.13.0
httptools==0.5.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
Mako==1.2.2
MarkupSafe==2.1.1
psycopg2==2.9.3
pydantic==1.10.2
python-decouple==3.6
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
starlette==0.20.4
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
zipp==3.8.1
